requirements.txt
imagemounter==3.1.0
pyaml==20.4.0
python-magic==0.4.22
python-vagrant==0.5.15
PyYAML==5.4.1
termcolor==1.1.0
